Dev/stam/spec 101 fixes (#8)
* tried copying some code from GetApplicationId but not getting any outputs yet...

* renaming queryspeckleobjectcompenent to be obsolete, reverted changes to the file

* created new guid and new file for new queryspeckleobjectcomponent

* updated description

* working on checking if the input is a nested value

* working on checking if input is a dictionary, returns correct value if not a dictionary

* created recursive function, corrected logic, still working on weird dictionary issue

* checked for IDictionary instead of being limited to just Dictionary

* trying to figure out the archivable dictionary situation

* fixed loop index to increment propertly

* removed top level check forcing users to provide Properties key

* added key to error message

* working on graft being turned on by default

* graft turned on by default

* added contents of getnestedprop to try catch

* returned null in catch, edited exception

* removed commented code

* updated warning message

* output item instead of list

* figured out how to return everything up to current index in error message

* added back _OBSOLETE querry object

* added hidden flag for Obsolete query component. Minor clean-up on usings an unused properties
